backend/requirements.txt
django
djangorestframework
django-cors-headers
anthropic
httpx
python-dotenv
